uploading images to ads

+++ app/src/main/java/com/example/rushan/RushWay/models/RWLocation.java
@@ -2,29 +2,29 @@ package com.example.rushan.RushWay.models;
 
 public class RWLocation {
 
-    private double lat;
-    private double lon;
+    private Long lat;
+    private Long lon;
     public RWLocation() {
     }
 
-    public RWLocation(double lat, double lon) {
+    public RWLocation(Long lat, Long lon) {
         this.lat = lat;
         this.lon = lon;
     }
 
-    public double getLat() {
+    public Long getLat() {
         return lat;
     }
 
-    public void setLat(double lat) {
+    public void setLat(Long lat) {
         this.lat = lat;
     }
 
-    public double getLon() {
+    public Long getLon() {
         return lon;
     }
 
-    public void setLon(double lon) {
+    public void setLon(Long lon) {
         this.lon = lon;
     }
 }
